requirements.txt
Django==5.2
python-decouple==3.8
django-cors-headers==4.3.1
whitenoise==6.6.0
djangorestframework==3.14.0
gunicorn==21.2.0
pymongo==4.6.1
openai>=1.12.0
google-generativeai==0.3.2
pytz==2024.1